refactor: split admin user list query helpers
